refactor: deconstruct monolithic InterviewSession page into hooks and presenter subcomponents
